refactor: Update transaction tables to typescript
- Update to typescript
- Use an intermediate named export
- Renamed components to not be plural

#399
