+++ src/containers/Accounts/AccountTransactionTable/index.ts
@@ -1,0 +1,3 @@
+import AccountTransactionTable from './AccountTransactionTable'
+
+export { AccountTransactionTable }
