+++ src/containers/Token/TokenTransactionTable/index.ts
@@ -1,0 +1,3 @@
+import TokenTransactionTable from './TokenTransactionTable'
+
+export { TokenTransactionTable }
